+++ Convex_hull_3/test/Convex_hull_3/test_extreme_points.cpp
@@ -212,12 +212,19 @@ void test_extreme_vertices(const char* fname)
     std::cerr << fname << " is not a valid off file.\n";
     exit(1);
   }  
-  CGAL::Extreme_points_traits_adapter_3<CGAL::Convex_hull_traits_3<K, Polyhedron_3, CGAL::Tag_true>
-      ,boost::property_map<Polyhedron_3, CGAL::vertex_point_t>::type >
-      traits(get(CGAL::vertex_point, P));
+  /*CGAL::Extreme_points_traits_adapter_3<
+      boost::property_map<Polyhedron_3, CGAL::vertex_point_t>::type
+      ,
+      CGAL::Convex_hull_traits_3<K, Polyhedron_3, CGAL::Tag_true>
+      >
+      traits(get(CGAL::vertex_point, P));*/
+  CGAL::Convex_hull_traits_3<K, Polyhedron_3, CGAL::Tag_true> traits;
+  boost::property_map<Polyhedron_3, CGAL::vertex_point_t>::type pmap = 
+      get(CGAL::vertex_point, P);
+  
   std::vector<boost::graph_traits<Polyhedron_3>::vertex_descriptor> verts;
   CGAL::extreme_points_3(vertices(P), std::back_inserter(verts) ,
-                   traits);
+                   CGAL::make_extreme_points_traits_adapter(pmap, traits));
 }
 
 int main()